requirements.txt
Flask==1.1.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
Jinja2==2.11.2
jstyleson==0.0.2
miniKanren==1.0.1
multidict==4.7.6
python-dotenv==0.15.0
pip==20.2.4